[Important] bug fix in AdaIn
I'm sorry to those who have used the code so far, but I found a bug while reading the code after 4 years.

The existing code had a fatal bug where the batch was split in two instead of the feature split in two.
(But what happened to the learning? There must have been a performance problem..)

This has been modified to divide normally using the chunk function.

Sorry again.